Added flexible parallel execution for the Folder Update.
Effectively speeds up import (depending on folder contents and cpus
avaliable) by about 30%, mostly improves hashing time and some
metadata collection operations that the filesystem can handle
more of a load.

Sqlite3 object no longer checks for same thread. Default sqlite3
thread safety mode is serialization: it is not as fast as if postrgresql
would be used but it is still thread safe for all operations.
https://www.sqlite.org/threadsafe.html

Futher improvements may be to create a processing job system to have ffmpeg
and PIL processes work on file objects asyncronously, but it would increase
the complexity substantially and not worth it at this time. Using cProfile
may be wise in order to perform further performance enhancements.